backend/ordersounds_mvp/requirements.txt
# Core PDF Processing
pdfplumber==0.11.0
pypdf==4.0.1
pdf2image==1.17.0

# Google Cloud Document AI
google-cloud-documentai==2.24.0
google-cloud-storage==2.14.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# String normalization
pycountry==23.12.11
python-Levenshtein==0.25.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
